safari-600.7-branch/LayoutTests/editing/secure-input/password-input-changed-type-expected.txt

Verify that changing a password input's type updates secure input state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


A password input is focused:
PASS testRunner.secureEventInputIsEnabled is true

After changing the type to text:
PASS testRunner.secureEventInputIsEnabled is false

After changing the type back to password:
PASS testRunner.secureEventInputIsEnabled is true
PASS successfullyParsed is true

TEST COMPLETE
